Ranks Renamed "premium" to "rank" Economy methods accepts Integers and Doubles Display names shows in player list Added player list header Money command can tell all accounts' balance grantAdvancement method is now in library

+++ plugin/src/main/kotlin/fr/pickaria/vue/reward/RewardCommand.kt
@@ -50,8 +50,8 @@ class RewardCommand : BaseCommand() {
 			throw ConditionFailedException(rewardConfig.cantPurchaseReward)
 		}
 
-		val totalKeys = (reward.keys * amount).toDouble()
-		val totalShards = (reward.shards * amount).toDouble()
+		val totalKeys = (reward.keys * amount)
+		val totalShards = (reward.shards * amount)
 
 		if (!sender.has(Key, totalKeys) || !sender.has(Shard, totalShards)) {
 			throw ConditionFailedException(rewardConfig.notEnoughMoney)
